% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtemplate_x3p.R
\name{addtemplate_x3p}
\alias{addtemplate_x3p}
\title{Add/change xml meta information in x3p object}
\usage{
addtemplate_x3p(x3p, template)
}
\arguments{
\item{x3p}{x3p object}

\item{template}{file path to xml template, use NULL for in-built package template}
}
\description{
Use the specified template to overwrite the general info in the x3p object (and structure of the feature info, if needed).
}
