% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lines_x3p.R
\name{calculate_spacing}
\alias{calculate_spacing}
\title{Calculate grid spacing}
\usage{
calculate_spacing(x3p, spaces, axis = "y")
}
\arguments{
\item{x3p}{x3p file}

\item{spaces}{space between grid lines}

\item{axis}{axis to calculate, as character}
}
\value{
vector of line locations
}
\description{
Helper function, not exported.
}
