% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_write.R
\name{convert_raster_to_png}
\alias{convert_raster_to_png}
\title{Helper function}
\usage{
convert_raster_to_png(raster)
}
\arguments{
\item{raster}{raster image}
}
\value{
png object
}
\description{
Helper function
}
