% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dim.x3p}
\alias{dim.x3p}
\title{Dimension of an x3p object}
\usage{
\method{dim}{x3p}(x)
}
\arguments{
\item{x}{an x3p object}
}
\value{
a vector of length three. The first two integers contain width and height of the surface scan, the third dimension is either 1 or 2, depending on whether there is a mask (2) or not (1).
}
\description{
\code{dim.x3p} expands the generic dim method for x3p objects and returns (mainly) the dimensions of the surface matrix.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
dim(logo)
}
