% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_x3p.R
\name{fortify_x3p}
\alias{fortify_x3p}
\title{Convert an x3p file into a data frame}
\usage{
fortify_x3p(x3p)
}
\arguments{
\item{x3p}{a file in x3p format as returned by function read_x3p}
}
\value{
data frame with variables x, y, and value and meta function in attribute
}
\description{
An x3p file consists of a list with meta info and a 2d matrix with scan depths. 
fortify turns the matrix into a data frame, using the parameters of the header as necessary.
}
\examples{
data(br411)
br411_fort <- fortify_x3p(br411)
head(br411_fort)
}
