% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_x3p.R
\name{head.x3p}
\alias{head.x3p}
\title{Show meta information of an x3p file}
\usage{
\method{head}{x3p}(x, n = 6L, ...)
}
\arguments{
\item{x}{x3p object}

\item{n}{number of rows/columns of the matrix}

\item{...}{extra parameters passed to head.matrix()}
}
\description{
\code{head.x3p} expands the generic head method for x3p objects. It gives a summary of the most relevant 3p meta information and returns the object invisibly.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
head(logo)
}
