% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image.x3p}
\alias{image.x3p}
\title{Raster image of an x3p surface}
\usage{
\method{image}{x3p}(x, ...)
}
\arguments{
\item{x}{an x3p object}

\item{...}{parameters passed into image}
}
\description{
This image function creates a quick raster image to show the surface of an x3p file.
Due to some inconsistency in where the origin (0,0) is supposed to be (top left or bottom left)  image functions from different packages will result in different images.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
image(logo)
}
