% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_x3p.R
\name{image_x3p_grid}
\alias{image_x3p_grid}
\title{Plot x3p object as an image}
\usage{
image_x3p_grid(x3p, file = NULL, col = "#cd7f32", spaces = 50,
  gridParam = list(color = "#e6bf98", radius = 5), size = c(750, 250),
  zoom = 0.35, multiply = 5, useNULL = FALSE,
  closeRGL = !is.null(file), ...)
}
\arguments{
\item{x3p}{x3p object}

\item{file}{file name for saving, if file is NULL the opengl device stays open.
The file extension determines the type of output. Possible extensions are png, stl (suitable for 3d printing), or svg.}

\item{col}{color specification}

\item{spaces}{space between grid lines, doubled for x}

\item{gridParam}{list with named components, consisting of parameters for showing crosscuts:  color and radius for crosscut region}

\item{size}{vector of width and height}

\item{zoom}{numeric value indicating the amount of zoom}

\item{multiply}{exaggerate the relief by factor multiply}

\item{useNULL}{logical value. Do you want a new rgl window to be opened? Use useNull = FALSE (default) in interactive mode and useNULL = TRUE if the image is opened as part of a predefined space such as an image container in an html site.}

\item{closeRGL}{logical value. Should the RGL window be closed by the routine?}

\item{...}{not used}
}
\description{
Plot x3p object as an image
}
\examples{
\dontrun{
logo <- read_x3p(system.file("csafe-logo.x3p", package="x3ptools"))
image_x3p(logo, file = "logo.png", crosscut = 50)
}
}
