% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_x3p.R
\name{interpolate_x3p}
\alias{interpolate_x3p}
\title{Interpolate from an x3p object}
\usage{
interpolate_x3p(x3p, resx = 1e-06, resy = resx)
}
\arguments{
\item{x3p}{x3p object}

\item{resx}{numeric value specifying the new resolution for the x axis}

\item{resy}{numeric value specifying the new resolution for the y axis}
}
\value{
interpolated x3p object
}
\description{
Interpolate from an x3p object
}
