% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lea}
\alias{lea}
\title{Subsampled scan of a land-engraved area}
\format{
\code{x3p} object
}
\usage{
lea
}
\description{
LEAs (land-engraved areas) are created on the outside of a bullet during
the firing process. Depending on the rifling inside the barrel, multiple lands
exist for each barrel. Striation marks in these land engraved areas are used
in forensic labs to determine whether two bullets were fired from the same firearm.
}
\examples{
data(lea)
image(lea)
if (interactive()) x3p_image(lea)
}
\keyword{datasets}
