% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NIST.R
\name{NISTstudy_download}
\alias{NISTstudy_download}
\title{Download data from an entire study}
\usage{
NISTstudy_download(study_link, directory, mirrorFileStructure = T)
}
\arguments{
\item{study_link}{Link to study page, or id value
(https://tsapps.nist.gov/NRBTD/Studies/Studies/Details/<study-id-value>)}

\item{directory}{Directory to download the data into.}

\item{mirrorFileStructure}{Should separate folders be created for each
firearm and bullet? This may result in highly nested data structure.
Setting this to FALSE may result in metadata being overwritten.}
}
\description{
Use this function to download all of the data from a single study. Locate
the study using the search page (\url{https://tsapps.nist.gov/NRBTD/Studies/Studies}).
}
\examples{
\dontrun{
url <- "https://tsapps.nist.gov/NRBTD/Studies/Studies/Details/"
studyID <- "c09aaa86-5d60-4acb-9031-46dad2c0ad32"
fullurl <- paste0(url, studyID)
NISTstudy_download(fullurl, 
                   file.path("data"), mirrorFileStructure = T)
NISTstudy_download(studyID, file.path("data"), mirrorFileStructure = T)                    
}
}
