% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NIST.R
\name{NRBTD_download}
\alias{NRBTD_download}
\title{Download data from an entire study}
\usage{
NRBTD_download(study_link, directory, mirrorFileStructure = T,
  maxFiles = 0, quiet = T)
}
\arguments{
\item{study_link}{Link to study page, or id value
(https://tsapps.nist.gov/NRBTD/Studies/Studies/Details/<study-id-value>)}

\item{directory}{Directory to download the data into.}

\item{mirrorFileStructure}{Should separate folders be created for each
firearm and bullet? This may result in highly nested data structure.
Setting this to FALSE may result in metadata being overwritten.}

\item{maxFiles}{maximum number of x3p files to download. 0 indicates that all x3p files should be downloaded.}

\item{quiet}{quiet download of files}
}
\description{
Use this function to download all of the data from a single study. Locate
the study using the search page (\url{https://tsapps.nist.gov/NRBTD/Studies/Studies}).
}
\section{Note}{

To instead download enough data to use for package exploration, use the
\code{\link{NRBTDsample_download}} function.
}

\examples{
\dontrun{
url <- "https://tsapps.nist.gov/NRBTD/Studies/Studies/Details/"
studyID <- "c09aaa86-5d60-4acb-9031-46dad2c0ad32"
fullurl <- paste0(url, studyID)
NRBTD_download(fullurl, 
                   file.path("data"), mirrorFileStructure = T)
NRBTD_download(studyID, file.path("data"), mirrorFileStructure = T)                    
}
}
\seealso{
NRBTDsample_download
}
