% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NIST.R
\name{NRBTDsample_download}
\alias{NRBTDsample_download}
\title{Download data from two NIST bullets to use as sample data}
\usage{
NRBTDsample_download(directory, maxFiles = 0, quiet = T)
}
\arguments{
\item{directory}{Location to save the files}

\item{maxFiles}{maximum number of x3p files to download. 0 indicates that all x3p files should be downloaded.}

\item{quiet}{quiet download of files}
}
\description{
Bullets downloaded are from the Hamby series of studies, using 9mm
ammunition. This function downloads two bullets from the same barrel
(Barrel 1); each bullet has 6 x3p files associated with it - one for each
land. The x3p files are placed in folders corresponding to each bullet.
Data from additional studies can be found at
\url{https://tsapps.nist.gov/NRBTD/Studies/Search} and can be downloaded with
the \code{\link{NRBTD_download}} function.
}
\examples{
\dontrun{
NRBTDsample_download(file.path("data"))                  
}
}
\seealso{
NRBTD_download
}
