% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_x3p.R
\name{print.x3p}
\alias{print.x3p}
\title{Show meta information of an x3p file}
\usage{
\method{print}{x3p}(x, ...)
}
\arguments{
\item{x}{x3p object}

\item{...}{ignored}
}
\description{
\code{print.x3p} expands the generic print method for x3p objects. It gives a summary of the most relevant 3p meta information and returns the object invisibly.
}
\examples{
logo <- read_x3p(system.file("csafe-logo.x3p", package="x3ptools"))
print(logo)
}
