% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_x3p.R
\name{read_x3p}
\alias{read_x3p}
\title{Read an x3p file into an x3p object}
\usage{
read_x3p(file, size = NA, quiet = T)
}
\arguments{
\item{file}{The file path to the x3p file, or an url to an x3p file}

\item{size}{size in bytes to use for reading the binary file. If not specified, default is used.}

\item{quiet}{for url downloads, show download progress?}
}
\value{
x3p object consisting of a list of the surface matrix and the four records as specified in the ISO standard
}
\description{
Read an x3p file into an x3p object
}
\examples{
logo <- read_x3p(system.file("csafe-logo.x3p", package="x3ptools"))
}
