% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_x3p.R
\name{sample_x3p}
\alias{sample_x3p}
\title{Sampling from an x3p object}
\usage{
sample_x3p(x3p, m = 2)
}
\arguments{
\item{x3p}{x3p object}

\item{m}{integer value - every mth value is included in the sample}
}
\value{
downsampled x3p object
}
\description{
Sampling from an x3p object
}
