% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_x3p.R
\name{stl_to_x3p}
\alias{stl_to_x3p}
\title{Convert an STL file to an x3p file}
\usage{
stl_to_x3p(stl)
}
\arguments{
\item{stl}{STL file object or path to the file}
}
\value{
x3p object
}
\description{
STL files describe 3d objects as mesh objects. Here, we assume that the 3d object consists of a 3d surface on the top of a 'box'.
We further assume, that each node of the STL file describes the x-y location of an actual measurement, and that measurements are taken on
a regular 2d grid. These measurements are then converted into the surface matrix of an x3p object.
The resolution is derived from the distance between consecutive x and y nodes.
}
\examples{
\dontrun{
# the website https://touchterrain.geol.iastate.edu/ allows a download
# of a 3d printable terrain model, that should be very similar to 
# the GelSight format.
gc <- rgl::readSTL("~/Downloads/671743070119/NED_-112.16_36.14_tile_1_1.STL", plot=FALSE)
x3p <- stl_to_x3p(gc)
stl_to_x3p("~/Downloads/671743070119/NED_-112.16_36.14_tile_1_1.STL")
}
}
