% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_x3p.R
\name{transpose_x3p}
\alias{transpose_x3p}
\title{Transpose an x3p object}
\usage{
transpose_x3p(x3p)
}
\arguments{
\item{x3p}{x3p object}
}
\description{
Transpose the surface matrix of an x3p object. Also adjust meta information.
}
