% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_x3p.R
\name{unfortify_x3p}
\alias{unfortify_x3p}
\title{Convert  a data frame into an x3p file}
\usage{
unfortify_x3p(dframe)
}
\arguments{
\item{dframe}{data frame}
}
\value{
x3p object
}
\description{
Convert  a data frame into an x3p file
}
