% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wire}
\alias{wire}
\title{Subsampled wire cut scan}
\format{
\code{x3p} object
}
\usage{
wire
}
\description{
An example part of a wire cut in \code{x3p} format. The wire cut is part of a CSAFE study
involving 1.5 mm Aluminium wires cut by Kaiweet wire-cutters.
}
\keyword{datasets}
