% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_x3p.R
\name{write_x3p}
\alias{write_x3p}
\title{Write an x3p object to a file}
\usage{
write_x3p(x3p, file)
}
\arguments{
\item{x3p}{x3p object}

\item{file}{path to where the file should be written}
}
\description{
Write an x3p object to a file
}
