% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_x3p.R
\name{x3p_add_annotation}
\alias{x3p_add_annotation}
\title{Add annotations to an x3p object}
\usage{
x3p_add_annotation(x3p, color, annotation)
}
\arguments{
\item{x3p}{x3p object}

\item{color}{name or hex value of color}

\item{annotation}{character value describing the region}
}
\value{
x3p object with the added annotations
}
\description{
Add annotations to an x3p object
}
