% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_add_mask_layer.R
\name{x3p_add_mask_layer}
\alias{x3p_add_mask_layer}
\title{Add a layer to the mask}
\usage{
x3p_add_mask_layer(x3p, mask, color = "red", annotation = "")
}
\arguments{
\item{x3p}{x3p object}

\item{mask}{logical matrix of the same dimension as the surface matrix. Values of TRUE are assumed to be added in the mask, values of FALSE are being ignored.}

\item{color}{name or hex value of color}

\item{annotation}{character value describing the region}
}
\value{
x3p object with changed mask
}
\description{
Add a region  a mask.  The region is specfied as TRUE values in a matrix of the same dimensions as the existing mask. In case no mask exists, one is created.
}
