% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_bin_strips.R
\name{x3p_bin_stripes}
\alias{x3p_bin_stripes}
\title{Add colored stripes of the surface gradient to the mask of an x3p object}
\usage{
x3p_bin_stripes(
  x3p,
  direction = "vertical",
  colors = rev(c("#b12819", "#d7301f", "#e16457", "#ffffff", "#5186a2", "#175d82",
    "#134D6B")),
  freqs = c(0, 0.05, 0.1, 0.3, 0.7, 0.9, 0.95, 1)
)
}
\arguments{
\item{x3p}{object containing a 3d topographic surface}

\item{direction}{in which the stripes are created: \code{vertical} or \code{horizontal}.}

\item{colors}{vector of colors}

\item{freqs}{vector of values corresponding to color frequency (turned into quantiles of the differenced values)}
}
\value{
x3p object with mask colored by discretized surface gradient
}
\description{
Apply gradient-based color shading to the mask of a 3d topographic surface.
gradients are determined empirically based on row- or column-wise differenced
sequential surface values. The \code{direction} parameter determines the direction
differencing. If direction is "vertical", columns in the surface matrix are
differenced to identify whether 'vertical' stripes exist.
}
\examples{
\dontrun{
x3p <- x3p_read("~/Documents/CSAFE/Wirecutter/data/Aluminum Wires renamed/T1AW-LI-R1-B01.x3p")
x3p_image(x3p, size = dim(x3p$surface.matrix)/5, zoom=0.8)
x3p_with <- x3p \%>\% x3p_bin_stripes(direction="vertical")
x3p_image(x3p_with \%>\% x3p_sample(5), size = dim(x3p$surface.matrix)/5, zoom=0.8)
# sampling first is better at preserving the intended color frequency
x3p_with2 <- x3p \%>\% x3p_sample(5) \%>\% x3p_bin_stripes(direction="vertical")
x3p_image(x3p_with2, size = dim(x3p$surface.matrix)/5, zoom=0.8)
x3p_with3 <- x3p \%>\% x3p_average(5) \%>\% x3p_bin_stripes(direction="vertical")
x3p_image(x3p_with3, size = dim(x3p$surface.matrix)/5, zoom=0.8)

lea <- x3p_read("~/Documents/Data/Houston Persistence/Barrel A/Bullet 11/Houston 1 - Barrel A - Bullet 11 - Land 1 - Sneox1 - 20x - auto light left image + 20\% x10 - threshold 2 - resolution 4 - Mark Mosher.x3p")
lea \%>\% x3p_sample(m=5) \%>\% x3p_bin_stripes() \%>\% x3p_image()
lea \%>\% x3p_sample(m=5) \%>\% x3p_bin_stripes(freqs = c(0, 0.05, 0.1, 0.2,0.8, 0.9, 0.95, 1)) \%>\% x3p_image()
}
}
