% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_x3p.R
\name{x3p_delete_mask}
\alias{x3p_delete_mask}
\title{Delete mask from an x3p object}
\usage{
x3p_delete_mask(x3p)
}
\arguments{
\item{x3p}{x3p object}
}
\value{
x3p object without the mask
}
\description{
Deletes mask and its annotations from an x3p file.
}
