% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_select.R
\name{x3p_extract_profile}
\alias{x3p_extract_profile}
\title{Interactively select a line on the active rgl device}
\usage{
x3p_extract_profile(x3p, col = "#FF0000", update = TRUE, line_out = TRUE)
}
\arguments{
\item{x3p}{x3p file}

\item{col}{character value of the selection color}

\item{update}{boolean value, whether the rgl window should be updated to show the selected circle}

\item{line_out}{boolean enhavce result by a data frame of the line? Note that variable x indicates the direction from first click (x=0) to
the second click (max x). The values of x in the result are in the same units as the original x3p.}
}
\value{
x3p file with identified in mask enhanced by a dataframe of the line segment (line_df).
}
\description{
Use the right mouse button (double-finger on track pad) to identify to points on the
active rgl device (this rgl device should show the surface of an x3p object). These points define the beginning and end  of a line segment.
The line segment is drawn on the mask of the x3p object.
}
