% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_select.R
\name{x3p_fuzzyselect}
\alias{x3p_fuzzyselect}
\title{Interactive selection of region of interest}
\usage{
x3p_fuzzyselect(x3p, col = "#FF0000", mad = 5, type = "plane", update = TRUE)
}
\arguments{
\item{x3p}{x3p file}

\item{col}{character value of the selection color}

\item{mad}{scalar}

\item{type}{only "plane" is implemented at the moment}

\item{update}{boolean value, whether the rgl window should be updated to show the selected rectangle}
}
\value{
x3p file with updated mask
}
\description{
Interactive selection of region of interest
}
