% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{x3p_get_scale}
\alias{x3p_get_scale}
\title{Check resolution of a scan}
\usage{
x3p_get_scale(x3p)
}
\arguments{
\item{x3p}{object}
}
\value{
numeric value of resolution per pixel
}
\description{
Scans in x3p format capture 3d topogographic surfaces. According to ISO standard
ISO5436 - 2000 scans are supposed to be captured in meters. For microscopic images
capture in meters might be impractical.
}
