% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{x3p_m_to_mum}
\alias{x3p_m_to_mum}
\title{Convert x3p header information to microns from meters}
\usage{
x3p_m_to_mum(x3p)
}
\arguments{
\item{x3p}{x3p file with header information in meters}
}
\value{
x3p with header information in microns
}
\description{
ISO standard 5436_2 asks for specification of values in meters.
For topographic surfaces collected by microscopes values in microns are
more readable. Besides scaling the values in the surface matrix, corresponding
increments are changed to microns as well.
}
