% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_mask_legend.R
\name{x3p_mask_legend}
\alias{x3p_mask_legend}
\title{Get legend for mask colors}
\usage{
x3p_mask_legend(x3p)
}
\arguments{
\item{x3p}{x3p object (with a mask)}
}
\value{
named vector of colors, names show annotations
}
\description{
Get legend for mask colors
}
