% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_mask_quantile.R
\name{x3p_mask_quantile}
\alias{x3p_mask_quantile}
\title{Draw a quantile region on the mask}
\usage{
x3p_mask_quantile(
  x3p,
  quantiles = c(0.25, 0.75),
  color = "red",
  annotation = "quantile-region"
)
}
\arguments{
\item{x3p}{x3p object}

\item{quantiles}{vector of quantiles between which surface matrix values are included in the mask}

\item{color}{name or hex value of color}

\item{annotation}{character value describing the region}
}
\value{
x3p object with changed mask
}
\description{
For each x value of the surface matrix add a region to the mask of an x3p object corresponding to the area between two specified
quantiles.
}
