% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_show_xml.R
\name{x3p_paste_comment}
\alias{x3p_paste_comment}
\title{Append a comment to the current x3p file}
\usage{
x3p_paste_comment(x3p, comment, sep = ", ")
}
\arguments{
\item{x3p}{x3p object}

\item{comment}{character value with the appendage}
}
\value{
x3p object with expanded comment
}
\description{
Append information to the current comment of the x3p object. If there is no comment, a comment will be included in Record2 (general info)
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
logo \%>\% x3p_paste_comment("*** added to the current comment ***")
}
