% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_read_dat.R
\name{x3p_read_dat}
\alias{x3p_read_dat}
\title{Read data from an x-y-z file}
\usage{
x3p_read_dat(dat, delim = " ", col_names = FALSE)
}
\arguments{
\item{dat}{path to the x-y-z file}

\item{delim}{character determining delimiter}
}
\value{
x3p object
}
\description{
Read data from an x-y-z file
}
