% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_read_folder.R
\name{x3p_read_folder}
\alias{x3p_read_folder}
\title{Read all x3p files in a folder}
\usage{
x3p_read_folder(path)
}
\arguments{
\item{path}{character string to a folder}
}
\value{
tibble with two variables: source contains the file path to the x3p file, and x3p is a list variable of x3p objects.
}
\description{
This function is provided with a path to a folder, it will search recursively for all x3p files within the folder, read the x3p files, and return the results as part of a tibble dataset.
}
\examples{
# example code
path <- dirname(system.file("csafe-logo.x3p", package="x3ptools"))
x3p_read_folder(path) # show all x3p files installed with x3ptools
}
