% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_read_dat.R
\name{x3p_read_plux}
\alias{x3p_read_plux}
\title{Read information from plux file}
\usage{
x3p_read_plux(plux)
}
\arguments{
\item{plux}{path to plux file}
}
\value{
xml of general information as stored in the plux file
}
\description{
Read information from plux file
}
