% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_rotate.R
\name{x3p_rotate}
\alias{x3p_rotate}
\alias{rotate_x3p}
\title{Rotate an x3p object}
\usage{
x3p_rotate(x3p, angle)

rotate_x3p(x3p, angle)
}
\arguments{
\item{x3p}{x3p object}

\item{angle}{rotate clockwise by angle degrees given as 30, 60, 90 degree.}
}
\description{
Rotate the surface matrix of an x3p object. Also adjust meta information.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
dim(logo$surface.matrix)
\dontrun{
x3p_image(logo)
}
# rotate the image by 60 degrees counter-clockwise:
logo60 <- x3p_rotate(logo, 60)
dim(logo60$surface.matrix)
\dontrun{
x3p_image(logo60, zoom = 1.2)
}
}
