% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_rotate_xy.R
\name{x3p_rotate_xy}
\alias{x3p_rotate_xy}
\title{Rotate an x3p object}
\usage{
x3p_rotate_xy(x3p, theta)
}
\arguments{
\item{x3p}{x3p object}

\item{angle}{rotate clockwise by angle degrees given as 30, 60, 90 degree.}
}
\description{
Rotate the surface matrix of an x3p object. Also adjust meta information.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
dim(logo$surface.matrix)
\dontrun{
x3p_image(logo)
}
# rotate the image by 60 degrees counter-clockwise:
logo60 <- x3p_rotate_xy(logo, 60)
dim(logo60$surface.matrix)
\dontrun{
x3p_image(logo60)
}
}
