% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_select.R
\name{x3p_select}
\alias{x3p_select}
\title{Draw rectangle on the surface of an x3p file using rgl}
\usage{
x3p_select(x3p, col = "#FF0000", update = TRUE)
}
\arguments{
\item{x3p}{x3p file}

\item{col}{character value of the selection color}

\item{update}{boolean value, whether the rgl window should be updated to show the selected rectangle}
}
\value{
x3p file with selection in mask
}
\description{
Draw rectangle on the surface of an x3p file using rgl
}
