% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_x3p.R
\name{x3p_snapshot}
\alias{x3p_snapshot}
\title{Take a snapshot of the active rgl device and save in a file}
\usage{
x3p_snapshot(file)
}
\arguments{
\item{file}{file name for saving.
The file extension determines the type of output. Possible extensions are png, stl (suitable for 3d printing), or svg.}
}
\description{
Make a snapshot of the current rgl device and save it to file. Options for file formats are png, svg, and stl (for 3d printing).
}
