% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_transpose.R
\name{x3p_transpose}
\alias{x3p_transpose}
\alias{transpose_x3p}
\title{Transpose an x3p object}
\usage{
x3p_transpose(x3p)

transpose_x3p(x3p)
}
\arguments{
\item{x3p}{x3p object}
}
\description{
Transpose the surface matrix of an x3p object. Also adjust meta information.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
dim(logo$surface.matrix)
\dontrun{
x3p_image(logo)
}
#  transpose the image
logotp <- x3p_transpose(logo)
dim(logotp$surface.matrix)
\dontrun{
x3p_image(logotp)
}
}
