% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_x3p.R
\name{df_to_x3p}
\alias{df_to_x3p}
\title{Convert  a data frame into an x3p file}
\usage{
df_to_x3p(dframe)
}
\arguments{
\item{dframe}{data frame. \code{dframe} must have the columns x, y, and value.}
}
\value{
x3p object
}
\description{
Convert  a data frame into an x3p file
}
