% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_x3p.R
\name{read_x3p}
\alias{read_x3p}
\title{Read an x3p file into an x3p object}
\usage{
read_x3p(file)
}
\arguments{
\item{file}{The file path to the x3p file}
}
\value{
x3p object consisting of a list of the surface matrix and the four records as specified in the ISO standard
}
\description{
Read an x3p file into an x3p object
}
\examples{
logo <- read_x3p(system.file("csafe-logo.x3p", package="x3ptools"))
}
