% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_x3p.R
\name{transpose_x3p}
\alias{transpose_x3p}
\title{Transpose an x3p object}
\usage{
transpose_x3p(x3p)
}
\arguments{
\item{x3p}{x3p object}
}
\description{
Transpose the surface matrix of an x3p object. Also adjust meta information.
}
\examples{
logo <- read_x3p(system.file("csafe-logo.x3p", package="x3ptools"))
dim(logo$surface.matrix)
\dontrun{
image_x3p(logo)
}
#  transpose the image
logotp <- transpose_x3p(logo)
dim(logotp$surface.matrix)
\dontrun{
image_x3p(logotp)
}
}
