% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_mask_legend.R
\name{x3p_add_legend}
\alias{x3p_add_legend}
\title{Add legend to active rgl object}
\usage{
x3p_add_legend(x3p, colors = NULL)
}
\arguments{
\item{x3p}{x3p object with a mask}

\item{colors}{named character vector of colors (in hex format by default), names contain annotations}
}
\description{
Add the legend for colors and annotations to the actuve rgl window.
}
\examples{
x3p <- read_x3p(system.file("sample-land.x3p", package="x3ptools"))
\dontrun{
image_x3p(x3p) # run when rgl can open window on the device
x3p_add_legend(x3p) # add legend
}
}
