% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_mask_legend.R
\name{x3p_darker}
\alias{x3p_darker}
\title{Darken active rgl object}
\usage{
x3p_darker()
}
\description{
Makes the currently active rgl object darker by removing a light source. Once all light sources are removed the object can not be any darker.
}
\examples{
x3p <- x3p_read(system.file("sample-land.x3p", package="x3ptools"))
\dontrun{
x3p_image(x3p) # run when rgl can open window on the device
x3p_darker() # remove a light source
}
}
