% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_extract_profile_segments.R
\name{x3p_extract_profile_segments}
\alias{x3p_extract_profile_segments}
\title{Extract profiles with multiple segments}
\usage{
x3p_extract_profile_segments(x3p, width, col, linewidth, verbose = TRUE)
}
\arguments{
\item{x3p}{object}

\item{width}{segment width}

\item{col}{color}

\item{linewidth}{integer value specifying the width for the profile}

\item{verbose}{logical}
}
\value{
x3p object with added lines
}
\description{
Extract profiles with multiple segments
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
logo <- logo \%>\% x3p_m_to_mum()
x3p_extract_profile_segments(logo, 850, col="#ffffff")
}
