% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_mask_legend.R
\name{x3p_lighter}
\alias{x3p_lighter}
\title{Lighten active rgl object}
\usage{
x3p_lighter()
}
\description{
Make the currently active rgl object lighter. Adds a light source. Up to eight light sources can be added. Alternatively, any rgl light source can be added (see \code{light3d}).
}
\examples{
x3p <- x3p_read(system.file("sample-land.x3p", package="x3ptools"))
\dontrun{
x3p_image(x3p) # run when rgl can open window on the device
x3p_lighter() # add a light source
}
}
